refactor : ConsoleCommand 클래스 view.command 패키지로 이동
